+++ src/test/java/com/intellij/plugins/haxe/ide/inlay/HaxeReturnTypeInlayTest.java
@@ -27,5 +27,10 @@ public class HaxeReturnTypeInlayTest extends HaxeInlayTestBase {
     doTest(hintsProvider);
   }
 
+  @Test
+  public void testReturnTypeGenerics() throws Exception {
+    doTest(hintsProvider);
+  }
+
 
 }